feat: add toastr notifications to product create/update

+++ src/app/components/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Product } from '../../common/product';
 import { ProductService } from '../../services/product.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-product-list',
@@ -27,12 +28,35 @@ export class ProductListComponent implements OnInit {
   }
 
   deleteProduct(productId: number): void {
-    this.productService.deleteProduct(productId).subscribe(
-      () => {
-        console.log(`Product with ID ${productId} deleted successfully.`);
-        this.listProducts(); // Refresh the product list after deletion
-      }
-    );  
+
+  Swal.fire({
+  title: "Está seguro que quiere eliminar el registro?",
+  text: "",
+  icon: "warning",
+  showCancelButton: true,
+  confirmButtonColor: "#3085d6",
+  cancelButtonColor: "#d33",
+  confirmButtonText: "Eliminar", 
+  cancelButtonText: "Cancelar", 
+}).then((result) => {
+  if (result.isConfirmed) {
+
+        this.productService.deleteProduct(productId).subscribe(
+          () => {
+            console.log(`Product with ID ${productId} deleted successfully.`);
+            this.listProducts(); // Refresh the product list after deletion
+          }
+        );  
+
+        Swal.fire({
+          title: "Producto",
+          text: "Producto Eliminado.",
+          icon: "success"
+        });
+       }
+     });
+
+    
   }
 
 }
